update navbar to display account name based on the user name or user login, change user service to skip elasticsearch repo when it's disabled in configuration

+++ src/main/java/com/sevago/mpc/service/UserService.java
@@ -1,5 +1,6 @@
 package com.sevago.mpc.service;
 
+import com.sevago.mpc.config.ApplicationProperties;
 import com.sevago.mpc.domain.Authority;
 import com.sevago.mpc.domain.User;
 import com.sevago.mpc.repository.AuthorityRepository;
@@ -10,7 +11,6 @@ import com.sevago.mpc.security.AuthoritiesConstants;
 import com.sevago.mpc.security.SecurityUtils;
 import com.sevago.mpc.service.util.RandomUtil;
 import com.sevago.mpc.service.dto.UserDTO;
-import com.sevago.mpc.web.rest.vm.ManagedUserVM;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -50,13 +50,16 @@ public class UserService {
 
     private final CacheManager cacheManager;
 
-    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder, SocialService socialService, UserSearchRepository userSearchRepository, AuthorityRepository authorityRepository, CacheManager cacheManager) {
+    private final ApplicationProperties applicationProperties;
+
+    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder, SocialService socialService, UserSearchRepository userSearchRepository, AuthorityRepository authorityRepository, CacheManager cacheManager, ApplicationProperties applicationProperties) {
         this.userRepository = userRepository;
         this.passwordEncoder = passwordEncoder;
         this.socialService = socialService;
         this.userSearchRepository = userSearchRepository;
         this.authorityRepository = authorityRepository;
         this.cacheManager = cacheManager;
+        this.applicationProperties = applicationProperties;
     }
 
     public Optional<User> activateRegistration(String key) {
@@ -66,7 +69,9 @@ public class UserService {
                 // activate given user for the registration key.
                 user.setActivated(true);
                 user.setActivationKey(null);
-                userSearchRepository.save(user);
+                if (applicationProperties.getElasticsearch().isEnabled()) {
+                    userSearchRepository.save(user);
+                }
                 cacheManager.getCache(USERS_CACHE).evict(user.getLogin());
                 log.debug("Activated user: {}", user);
                 return user;
@@ -119,7 +124,9 @@ public class UserService {
         authorities.add(authority);
         newUser.setAuthorities(authorities);
         userRepository.save(newUser);
-        userSearchRepository.save(newUser);
+        if (applicationProperties.getElasticsearch().isEnabled()) {
+            userSearchRepository.save(newUser);
+        }
         log.debug("Created Information for User: {}", newUser);
         return newUser;
     }
@@ -148,7 +155,9 @@ public class UserService {
         user.setResetDate(Instant.now());
         user.setActivated(true);
         userRepository.save(user);
-        userSearchRepository.save(user);
+        if (applicationProperties.getElasticsearch().isEnabled()) {
+            userSearchRepository.save(user);
+        }
         log.debug("Created Information for User: {}", user);
         return user;
     }
@@ -171,7 +180,9 @@ public class UserService {
                 user.setEmail(email);
                 user.setLangKey(langKey);
                 user.setImageUrl(imageUrl);
-                userSearchRepository.save(user);
+                if (applicationProperties.getElasticsearch().isEnabled()) {
+                    userSearchRepository.save(user);
+                }
                 cacheManager.getCache(USERS_CACHE).evict(user.getLogin());
                 log.debug("Changed Information for User: {}", user);
             });
@@ -199,7 +210,9 @@ public class UserService {
                 userDTO.getAuthorities().stream()
                     .map(authorityRepository::findOne)
                     .forEach(managedAuthorities::add);
-                userSearchRepository.save(user);
+                if (applicationProperties.getElasticsearch().isEnabled()) {
+                    userSearchRepository.save(user);
+                }
                 cacheManager.getCache(USERS_CACHE).evict(user.getLogin());
                 log.debug("Changed Information for User: {}", user);
                 return user;
@@ -211,7 +224,9 @@ public class UserService {
         userRepository.findOneByLogin(login).ifPresent(user -> {
             socialService.deleteUserSocialConnection(user.getLogin());
             userRepository.delete(user);
-            userSearchRepository.delete(user);
+            if (applicationProperties.getElasticsearch().isEnabled()) {
+                userSearchRepository.delete(user);
+            }
             cacheManager.getCache(USERS_CACHE).evict(login);
             log.debug("Deleted User: {}", user);
         });
@@ -259,7 +274,9 @@ public class UserService {
         for (User user : users) {
             log.debug("Deleting not activated user {}", user.getLogin());
             userRepository.delete(user);
-            userSearchRepository.delete(user);
+            if (applicationProperties.getElasticsearch().isEnabled()) {
+                userSearchRepository.delete(user);
+            }
             cacheManager.getCache(USERS_CACHE).evict(user.getLogin());
         }
     }
